Create carlys-clippers.py
You will be calculating some important metrics that Carly can use to plan out the operation of the business for the rest of the month.
There are 3 three lists:
- hairstyles: the names of the cuts offered at Carly’s Clippers.
- prices: the price of each hairstyle in the hairstyles list.
- last_week: the number of purchases for each hairstyle type in the last week.
Each index in hairstyles corresponds to an associated index in prices and last_week.

For example, The hairstyle "bouffant" has an associated price of 30 from the prices list, and was purchased 2 times in the last week as shown in the last_week list. Each of these elements are in the first index of their respective lists.